add simulation/reality part
