--- swarm/launch/follower_launch_file.launch.py
@@ -1,11 +1,0 @@
-from launch import LaunchDescription
-from launch_ros.actions import Node
-
-def generate_launch_description():
-    return LaunchDescription([
-        Node(
-            package='swarm',
-            executable='follower_node',
-            output='screen'
-        )
-    ])
