--- swarm/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/swarm
-[install]
-install_scripts=$base/lib/swarm
