--- swarm/setup.py
@@ -1,30 +1,0 @@
-from setuptools import find_packages, setup
-import os
-from glob import glob
-
-package_name = 'swarm'
-
-setup(
-    name=package_name,
-    version='0.0.0',
-    packages=find_packages(exclude=['test']),
-    data_files=[
-        ('share/ament_index/resource_index/packages',
-            ['resource/' + package_name]),
-        ('share/' + package_name, ['package.xml']),
-        (os.path.join('share', package_name), glob('launch/*.launch.py'))
-    ],
-    install_requires=['setuptools'],
-    zip_safe=True,
-    maintainer='user',
-    maintainer_email='user@todo.todo',
-    description='TODO: Package description',
-    license='TODO: License declaration',
-    tests_require=['pytest'],
-    entry_points={
-        'console_scripts': [
-            'leader_node = swarm.leader:main',
-            'follower_node = swarm.follower:main',
-        ],
-    },
-)
