+++ wk12 - swarm_simulation/swarm/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/swarm
+[install]
+install_scripts=$base/lib/swarm
