+++ wk13 - swarm_reality/swarm/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/swarm
+[install]
+install_scripts=$base/lib/swarm
